requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.0.8
django-crispy-forms==1.9.1
django-emoji==2.2.2
django-formtools==2.2
django-import-export==2.2.0
django-pure-pagination==0.3.0
django-reversion==3.0.7
djangorestframework==3.11.0
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.18.1
idna==2.10
jdcal==1.4.1
Markdown==3.2.2
MarkupPy==1.14
mysqlclient==2.0.1
odfpy==1.4.1
openpyxl==3.0.4
Pillow==7.2.0
Pygments==2.6.1
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
urllib3==1.25.9
xlrd==1.2.0
XlsxWriter==1.2.9
xlwt==1.3.0
